Introduce new @prairielearn/html-ejs package
