+++ packages/html-ejs/package.json
@@ -1,14 +1,19 @@
 {
-  "name": "@prairielearn/html",
-  "version": "1.0.2",
+  "name": "@prairielearn/html-ejs",
+  "version": "1.0.0",
   "main": "dist/index.js",
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch --preserveWatchOutput",
     "test": "mocha --no-config --require ts-node/register src/index.test.ts"
   },
+  "dependencies": {
+    "@prairielearn/html": "^1.0.0",
+    "ejs": "^3.1.6"
+  },
   "devDependencies": {
     "@prairielearn/tsconfig": "*",
+    "@types/ejs": "^3.1.0",
     "mocha": "^9.2.2",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
