+++ packages/html/README.md
@@ -40,26 +40,6 @@ console.log(html`
 `);
 ```
 
-### Using with EJS
-
-If you have an EJS partial that you'd like to use inside of an `html` tagged template literal, you can use the `renderEjs` helper:
-
-```html
-<!-- hello.ejs -->
-Hello, <%= name %>!
-```
-
-```js
-const { hmtl, renderEjs } = require('@prairielearn/html');
-
-console.log(
-  html`
-    <div>Hello, world!</div>
-    <div>${renderEjs(__filename, "<%- include('./hello'); %>", { name: 'Anjali' })}</div>
-  `.toString()
-);
-```
-
 ## Why not EJS?
 
 PrairieLearn used (and still uses) EJS to render most views. However, using a tagged template literal and pure JavaScript to render views has a number of advantages:
@@ -67,3 +47,5 @@ PrairieLearn used (and still uses) EJS to render most views. However, using a ta
 - Prettier will automatically format the content of any `html` tagged template literal; EJS does not have any automatic formatters.
 - Authoring views in pure JavaScript allows for easier and more explicit composition of components.
 - It's possible to use ESLint and TypeScript to type-check JavaScript views; EJS does not offer support for either.
+
+If you want to use existing EJS partials inside of an `html` tagged template literal, check out the `@prairielearn/html-ejs` package. EJS-related functionality is deliberately located in a separate package so that `@prairielearn/html` can be used in the browser, since the `ejs` package makes use of Node-only features.
