+++ packages/opentelemetry/tsconfig.json
@@ -1,16 +1,7 @@
 {
+  "extends": "@prairielearn/tsconfig",
   "compilerOptions": {
-    "allowJs": true,
-    "declaration": true,
-    "esModuleInterop": true,
     "outDir": "./dist",
-    "rootDir": "src/",
-    "sourceMap": true,
-    // This package will only be used server-side on Node 14+, so target the
-    // newest version of the ES spec that Node 14 supports.
-    "target": "ES2020",
-    // However, we don't yet make extensive use of ES Modules, so specifically
-    // compile `import`/`export` down to CommonJS.
-    "module": "CommonJS",
+    "rootDir": "./src",
   }
 }
