+++ packages/tsconfig/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@prairielearn/tsconfig",
+  "version": "0.0.0",
+  "private": true,
+  "main": "tsconfig.json",
+  "files": [
+    "tsconfig.json"
+  ]
+}
