+++ packages/tsconfig/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    "esModuleInterop": true,
+    "sourceMap": true,
+    "target": "ES2020",
+    // We don't yet make extensive use of ES Modules, so compile
+    // `import`/`export` down to CommonJS.
+    "module": "CommonJS",
+  }
+}
